telegrambot/requirements.txt
aiohttp
uvloop
ujson
aiogram
pillow
numpy
pandas
tensorflow
keras
opencv-python
torch
torchvision

!python -m pip install detectron2 -f \
  https://dl.fbaipublicfiles.com/detectron2/wheels/cu110/torch1.7/index.html
